pyproject.toml
[tool.poetry]
name = "float-project"
version = "0.1.0a0"
description = "A latent-thought based learning agent"
authors = ["Cherry Research <cherry.research@pm.me>"]
license = "AGPL-3.0-only"
packages = [
    { include = "app", from = "backend" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.34.3"
pydantic = "^2.11.7"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
cryptography = "^45.0.4"
celery = "^5.5.3"
redis = { version = "^6.2.0", optional = true }
python-dotenv = "^1.1.0"
httpx = "^0.28.1"
requests = "^2.32.4"
pandas = "^2.3.0"
networkx = "^3.1"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"
pyopenssl = "^25.1.0"
weaviate-client = "^4.9.2"
alembic = "^1.16.2"
beautifulsoup4 = "^4.13.4"
watchdog = "^6.0.0"
python-docx = "^1.2.0"
PyPDF2 = "^3.0.1"
sentence-transformers = "^5.0.0"
kokoro = { version = "^0.9.4", python = ">=3.11,<3.13" }
kittentts = "^0.1.3"
soundfile = "^0.12.1"
mcp = { version = "^1.12.2", extras = ["fastmcp"] }
langextract = "^1.0.2"
transformers = "^4.55.0"
huggingface-hub = { version = "^0.34.3", extras = ["hf-xet", "hf-transfer"] }
openai-harmony = "0.0.3"
icalendar = "^5.0.11"
python-dateutil = "^2.9.0"
pywebpush = "^1.14.0"
python-json-logger = { version = "^2.0.7", optional = true }
opentelemetry-exporter-otlp = { version = "^1.27.0", optional = true }
opentelemetry-instrumentation-fastapi = { version = "^0.48b0", optional = true }
opentelemetry-instrumentation-requests = { version = "^0.48b0", optional = true }
opentelemetry-instrumentation-logging = { version = "^0.48b0", optional = true }
prometheus-client = { version = "^0.20.0", optional = true }
flower = { version = "^2.0.1", optional = true }
scikit-learn = "^1.5.0"
Pillow = "^10.4.0"
hf-transfer = "^0.1.9"
chromadb = "^1.1.0"
mcp-server = "^0.1.4"
accelerate = "^1.10.1"
ipykernel = "^6.30.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pre-commit = "^4.2.0"

[tool.poetry.extras]
# Optional extras for background workers (Redis broker & Flower monitoring)
workers = ["redis", "flower"]
telemetry = [
    "python-json-logger",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-logging",
    "prometheus-client",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
float = "main:main"

[tool.isort]
profile = "black"
